C based minimisation
